requirements.txt
pyTelegramBotAPI
torch
torchvision
opencv-python-headless
numpy
Pillow